move cubesigner to a standalone file
